+++ buildSrc/src/main/kotlin/versions.kt
@@ -6,6 +6,7 @@ object vers {
 
     object asoft {
         val test = "1.1.10"
+        val expect = "0.0.10"
         val builders = "1.3.0"
         val persist = "0.2.0"
         val paging = "0.0.30"