+++ persist-keyvalue/persist-keyvalue-locally/src/jvmMain/kotlin/tz/co/asoft/LocalKeyValueStorage.kt
@@ -2,20 +2,36 @@ package tz.co.asoft
 
 import java.io.File
 
-class LocalKeyValueStorage(override val namespace: String = "", val file: File = File("$namespace.storage.json")) : KeyValueStorage {
+class LocalKeyValueStorage(override val namespace: String = "namespace", val file: File) : KeyValueStorage {
     init {
-        if(!file.exists()) {
-            file.createNewFile()
-        }
+        val parent = file.parentFile
+        if (!parent.exists()) parent.mkdirs()
+        if (!file.exists()) file.createNewFile()
+        require(namespace.isNotBlank()) { "Blank/Empty namespaces are not allowed" }
     }
-    val data = mutableMapOf<String, String>()
-    override fun clear() = data.clear()
-    override fun get(key: String): String? = data[key]
+
+    private val mapper = Mapper { prettyPrint = true }
+
+    private fun File.readMap() = try {
+        (mapper.decodeFromString(readText()) as? Map<String, String>) ?: mapOf()
+    } catch (_: Throwable) {
+        mapOf()
+    }
+
+    private fun File.write(map: Map<String, Any?>) = writeText(mapper.encodeToString(map))
+
+    override fun clear() = file.write(mapOf())
+
+    override fun get(key: String): String? = file.readMap()[key]
     override fun remove(key: String) {
-        data.remove(key)
+        val map = file.readMap().toMutableMap()
+        map.remove(key)
+        file.write(map)
     }
 
     override fun set(key: String, value: String) {
-        data[key] = value
+        val map = file.readMap().toMutableMap()
+        map[key] = value
+        file.write(map)
     }
 }